Move the ContactRegisterSettings to a new namespace
